working on actual blog page(other than bare rest operations) with jinja2

+++ backend/db/repository/blog.py
@@ -24,7 +24,9 @@ def list_blogs(db:Session):
 def update_blog_by_id(id:int, blog:UpdateBlog, db:Session, author_id:int=1):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {'error': f'Blog with id {id} not found'}
+    if not blog_in_db.author_id == author_id:
+        return {'error': 'You are not the author of this blog'}
     blog_in_db.title = blog.title
     blog_in_db.slug = blog.slug
     blog_in_db.content = blog.content
@@ -37,6 +39,8 @@ def delete_blog_by_id(id:int, db:Session, author_id:int):
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
         return {'error': f'Blog with id {id} not found'}
+    if not blog_in_db.first().author_id == author_id:
+        return {'error': 'You are not the author of this blog'}
     # blog_in_db.is_active = False
     # db.add(blog_in_db)
     blog_in_db.delete()
